Moved classes into their respective packages
