+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:theme="@style/AppTheme">
         <activity android:name=".ui.recipe.RecipeActivity"></activity>
         <activity android:name=".ui.recipe.NewRecipeActivity" />
-        <activity android:name=".RegisterNewUser" />
+        <activity android:name=".ui.login.RegisterNewUser" />
         <activity
             android:name=".ui.recipe.BMACRecipeListWindow"
             android:label="@string/title_activity_b_m_a_c_recipe_list_window"
