+++ app/src/main/java/com/example/bakemeacake/ui/login/RegisterNewUser.java
@@ -1,4 +1,4 @@
-package com.example.bakemeacake;
+package com.example.bakemeacake.ui.login;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -6,6 +6,7 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
+import com.example.bakemeacake.R;
 import com.example.bakemeacake.util.DatabaseHandler;
 
 public class RegisterNewUser extends AppCompatActivity {
