+++ app/src/main/res/layout/activity_register_new_user.xml
@@ -9,7 +9,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".RegisterNewUser">
+    tools:context=".ui.login.RegisterNewUser">
 
     <TextView
         android:id="@+id/textView_Username"
